ML/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Pillow
numpy
easyocr
opencv-python-headless
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv
google-generativeai
requests

